Refactor helpers and queries
