+++ backend/helpers/fileHelper.js
@@ -1,0 +1,27 @@
+const multer = require("multer");
+
+const fileHelper = {
+  storage: multer.diskStorage({
+    destination: (req, file, cb) => {
+      cb(null, "./images");
+    },
+    filename: (req, file, cb) => {
+      const uniqueSuffix = Date.now() + "-" + Math.round(Math.random() * 1e9);
+      cb(null, `${file.fieldname} - ${uniqueSuffix}`);
+    },
+  }),
+  fileFilter: (req, file, cb) => {
+    if (
+      file.mimetype == "image/png" ||
+      file.mimetype == "image/jpg" ||
+      file.mimetype == "image/jpeg"
+    ) {
+      cb(null, true);
+    } else {
+      cb(null, false);
+      return cb(new Error("Only .png, .jpg and .jpeg format allowed!"));
+    }
+  },
+};
+
+module.exports = fileHelper;
